backend/requirements.txt
autumn-py @ git+https://github.com/useautumn/autumn-py@9e09bbd34d49a41d739dc94fd9355783df65d178
aiohttp==3.12.13
starlette==0.47.1
fastapi==0.115.14
uvicorn==0.34.3